fix windows build

+++ src/Generator.hpp
@@ -25,12 +25,12 @@ namespace tulip::hook {
 		virtual Result<std::vector<uint8_t>> generateIntervener() = 0;
 
 		struct RelocateReturn {
-			size_t m_trampolineOffset;
-			size_t m_originalOffset;
+			uint64_t m_trampolineOffset;
+			uint64_t m_originalOffset;
 		};
 
 		virtual Result<> generateTrampoline(RelocateReturn offsets) = 0;
-		virtual Result<RelocateReturn> relocateOriginal(size_t target) = 0;
+		virtual Result<RelocateReturn> relocateOriginal(uint64_t target) = 0;
 
 		virtual std::string handlerString() = 0;
 		virtual std::string intervenerString() = 0;
