+++ src/platform/MacosGenerator.hpp
@@ -15,7 +15,7 @@ namespace tulip::hook {
 		Result<> generateHandler() override;
 		Result<std::vector<uint8_t>> generateIntervener() override;
 		Result<> generateTrampoline(RelocateReturn offsets) override;
-		Result<RelocateReturn> relocateOriginal(size_t target) override;
+		Result<RelocateReturn> relocateOriginal(uint64_t target) override;
 
 		std::string handlerString() override;
 		std::string intervenerString() override;
